Don't publish constraints from versions.props (#193)
Properly remove versions.props constraints from being published. This makes us now compatible with gradle metadata.

This is done by not configuring versions.props constraints on _any_ configuration that `apiElements` or `runtimeElements` extends from.

In order to maintain resolvability of the legacy configurations `compile`, `runtime` and `compileOnly`, we inject their versions at resolution time in the same way that nebula-dependency-recommender works.
We can't apply constraints on them anymore because they ancestors of the published configurations and so those constraints would be published.