Enforce lock constraints using shared platform (#557)
Manage lock file constraints using a single platform, reducing the number of constraints that have to be created.

This works in a similar way to [Java Platform Plugin](https://docs.gradle.org/current/userguide/java_platform_plugin.html). We can't just use that plugin directly, as it makes too many assumptions about the structure of the build where it's applied (e.g. needs its own project that isn't otherwise a java project, to hold the platform) that we can't make in GCV.