Add compatibility with Gradle configuration-on-demand
These didn't previously work well together because GCV wanted all
projects to be configured during the configuration of the project
(and in fact forced their configurations), while CoD configures
only a subset of projects. To help deal with this incompatibility,
GCV also had some direct checks telling you to disable CoD
entirely in order to run key operations.

Now, we only force the full set of projects to be configured when
using --write-locks, and otherwise allow commands like `./gradlew
:some-project:build` to only configure projects that we'd expect
to be configured by CoD.

This implementation is a bit of a hack, in that it makes some
conservative guesses about which projects to configure rather than
delegating to Gradle's "natural" configuration of those projects.
I first attempted the latter, but there are too many restrictions
on the timing of the various steps -- when dependencies can be
defined or resolved, when we can force a project to configure
itself, and so on. The good news is that the hackiness is confined
to a single function, and the known inaccuracies should result in
slower configuration rather than correctness errors.