add versions.list as clearer versions.lock diff target
Scanning diffs on versions.lock (e.g., reviewing an upgrade PR for any
unexpected changes) is tedious and error prone, as the constraint counts
and hashes change when there is no substantive change in the dependencies.
This adds an additional versions.list file alongside versions.lock that
excludes the constraint counts and hashes, making human reviews of
dependencies much faster and more reliable.